Add Scrobbles column to Tracks chart
Load the playcount for each track into the table. This is the total
number of times a track has been scrobbled and is scraped from the
page of each track. Sometimes this is a better measure of popularity
than the amount of listeners.

Last.fm has changed the HTML of their pages several times in the past.
Therefore, the code may need revising in the future. Previously, the
following lines were used:

    cell.load(url + " [itemprop=playCount]");
    cell.load(url + " [itemprop=interactionCount]");

A faster and more robust solution would be to use the interface
described at http://www.last.fm/api/show/track.getInfo, but this
requires an API key.